requirements.txt
appdirs==1.4.3
Django==3.1.7
Pillow==8.2.0
PyJWT==1.7.1
asgiref==3.3.1
django-cors-headers==3.7.0
djangorestframework==3.12.4
pip==21.0.1
psycopg2==2.8.6
python-dotenv==0.15.0
pytz==2021.1
setuptools==54.0.0
sqlparse==0.4.1